Create scraper.js module and move buildLinkList() there.
