+++ index.js
@@ -3,34 +3,12 @@ const fs = require('fs');
 const stat = fs.statSync;
 const path = require('path');
 const clear = require('clear');
-const Xray = require('x-ray');
 const parse = require('url-parse');
-const x = Xray();
 const yazl = require('yazl');
 const rimraf = require('rimraf');
 const isHtml = require('is-html');
 
-/**
- * @requires NPM:x-ray
- * @param {string} username Target freeCodeCamp username for scraping
- * @returns {promise} List of urls containing a solution query string
- */
-const buildLinkList = (username) => {
-    return new Promise((resolve, reject) => {
-        let linkList = [];
-        x(`https://freecodecamp.org/${username}`, ['a@href'])((err, data) => {
-            let links = data.toString().split(',');
-            for (let link of links) {
-                if (link.match(/\?solution=/)) {
-                    linkList.push(link);
-                } 
-            }
-            const linkListDeDuped = Array.from(new Set(linkList));
-            resolve(linkListDeDuped);
-        });
-    })
-};
-
+const scraper = require('./scraper');
 
 /**
  * Parses the URLs from the hrefs to obtain the challenge name and solution
@@ -75,12 +53,12 @@ const writeFile = (fileObject, count, camper) => {
  * Creates a zip file based on the contents of the solutions folder
  * @requires 'fs'
  * @requires NPM:yazl
- * @param {string} zipFileName The name of the output zipfile 
+ * @param {string} zipFileName The name of the output zipfile
  * @param {array} pathNames An array of all filenames to be zipped
  */
 const newArchive = (zipFileName, pathNames) => {
     const zipfile = new yazl.ZipFile();
-    
+
     pathNames.forEach(target => {
         const joinedTarget = path.join('./solutions/', target);
         const p = stat(joinedTarget);
@@ -91,9 +69,9 @@ const newArchive = (zipFileName, pathNames) => {
     zipfile.outputStream.pipe(fs.createWriteStream(zipFileName)).on("close", () => {
         console.log(`
         I zipped it real good!
-        
+
         You can collect your zipped archive from the fcc-scraper directory.
-        
+
         *If you found this useful, star the JacksonBates/fcc-scraper repo on GitHub*
         Twitter, Instagram and FCC forum: @jacksonbates
         `);
@@ -103,7 +81,7 @@ const newArchive = (zipFileName, pathNames) => {
 }
 
 const writeSolutions = (camper) => {
-    buildLinkList(camper)
+    scraper.buildLinkList(camper)
         .then((links) => {
             const count = links.length;
             for (let link of links) {
@@ -120,23 +98,23 @@ clear();
 rl.question('Please enter the username to scrape: ', (camper) => {
     writeSolutions(camper);
     console.log(`
-        ***          **             ***     
-       **            ***              **    
-      **             ****              **   
-     **              ****  **           **  
-    **              *****  **           *** 
-    **              **********           ** 
-    **             ***** *****           ** 
+        ***          **             ***
+       **            ***              **
+      **             ****              **
+     **              ****  **           **
+    **              *****  **           ***
+    **              **********           **
+    **             ***** *****           **
     **            *****  *****           ***
     **            ****   *****           ***
-    **            ***       *            ** 
-    **             **      **            ** 
-    ***             *                    ** 
-     **              *                  **  
-      **                               **   
-       **        ***************      **    
-        ***                         ***     
-          *                         *       
+    **            ***       *            **
+    **             **      **            **
+    ***             *                    **
+     **              *                  **
+      **                               **
+       **        ***************      **
+        ***                         ***
+          *                         *
     `);
     rl.close();
 });
